Adding a bunch of code to test examples with
There is a native C example that works – for reference and testing
Right now the rust example doesn't use much from the generated
ffi code, instead pokes at stuff itself. This is not ideal and
doesn't actually support that much functionality.

Ideally we want to wrap around that more. But also, replace large
chunks of the code with auto-generated proc-macro code (wip)